+++ src/main/steps/applicant1/address/manual/content.ts
@@ -3,7 +3,7 @@ import { FormContent, FormFields } from '../../../../app/form/Form';
 import {
   form as manualAddressForm,
   generateContent as manualAddressGenerateContent,
-} from '../../../common/components/manual-address';
+} from '../../../common/components/address-manual';
 
 const en = manualAddressContent => ({
   section: 'Primary applicant',
